requirements.txt
requests
colorama
prettytable
pillow
pytesseract
python-magic
PyPDF2
python-docx
pefile
